added new api endpoints for dxcenter page
