+++ server/server.js
@@ -21,6 +21,10 @@ sequelize
   })
   .done();
 
+require('./routes/dxsearch.js')(app, sequelize);
+require('./routes/dxdetails.js')(app, sequelize);
+require('./routes/dxtests.js')(app, sequelize);
+
 app.use(express.static('./dist'));
 app.set('views', path.join(__dirname, '../views'));
 
@@ -28,26 +32,6 @@ app.engine('html', require('ejs').renderFile);
 
 app.get('/', (req, res) => {
   res.render('index.html');
-});
-
-app.get('/searchdx', (req, res) => {
-  const inputString = req.query.q;
-  const searchString = inputString.concat('%');
-  sequelize
-    .query('SELECT id, name FROM diagnostic_center  WHERE `name` like :search_name', {
-      replacements: {
-        search_name: searchString,
-      },
-      type: sequelize.QueryTypes.SELECT,
-    })
-    .then((myTableRows) => {
-      res.end(JSON.stringify({ result: myTableRows }));
-    })
-    .catch((err) => {
-      console.log('SOMETHING WENT WRONG!', err);
-      res.end(JSON.stringify(err));
-    })
-    .done();
 });
 
 app.listen(3000, () => {
